CB-474 feat: divide pet info page and edit page
반려동물 정보 및 정보 수정 페이지 분리
- 새로운 디자인 적용
